MNT - M153532 - race condition in ContainsExpression
The problem is with the initialize call in the ContainsExpression method calling the setContainsId while another thread executing the same ContainsExpression initialize method, doing a getPCPathString on the same PCPath object.
So while a thread adds elements to the _actions list, another is iterating through it, causing CME.
The first call to setContainsId will set the _cid flag to true, so after that flag is set the _actions linkedlist is not updated anymore. (those we only get the error at the beginning of running a specific flow)

Moved the setContainsId logic from the ContainsExpression into the PCPath object under a sync block so that we do have two threads accessing the _actions list at the same time (one changing its contents and one iterating through them)

Also, created a ticket on openJPA:
https://issues.apache.org/jira/browse/OPENJPA-2610

Once that is fixed, we migth go for their fix if they decide on doing it some other way, as they know their code better.

git-svn-id: https://svn2.phx.axway.int/svngi5/svn/3rd_party_customized/repository/openjpa/branches/2.1.1@113 a3eec59e-6d41-4ad8-9ab7-c41e3df03bcb